dremote: reduce amount of signals

+++ src/arch/linux/async/signal.c
@@ -262,7 +262,7 @@ static void sigbreak(void *uc)
 {
   /* let CPUEMU decide what to do, as it can kick in for any backend */
   e_gen_sigalrm();
-  if (config.dpmi_remote && dpmi_pid)
+  if (config.dpmi_remote && dpmi_pid && in_rdpmi)
     kill(dpmi_pid, SIG_THREAD_NOTIFY);
 }
